+++ scripts/operation/start-trubudget.sh
@@ -26,6 +26,9 @@ Help()
     echo "                                  Available services: email-service, excel-export-service, storage-service"
     echo "  --no-log                        Disable logs of all docker-containers"
     echo "  --provision                     Start the provisioning"
+    echo "  --add-slave                     Add a slave-node that trys to connect to master-node"
+    echo "  --add-organization              Add a slave-node, slave-api, slave-frontend from a new Organization."
+    echo "                                  Needs to be approved by master-node"
     echo "  --prune                         Delete the multichain, document storage and email database (docker volume)"
     echo "  --help                          Print the help section"
     echo
@@ -46,7 +49,10 @@ IS_LOG_ENABLED=true
 WITH_PROVISIONING=false
 PRUNE_DATA=false
 COMPOSE_SERVICES=""
+ENABLED_SERVICES=""
+SLAVE_SERVICES=""
 IS_FULL=false
+HAS_SLAVE=false
 
 while [ "$1" != "" ]; do
     case $1 in
@@ -69,6 +75,21 @@ while [ "$1" != "" ]; do
             shift # past argument
         ;;
         
+        --add-slave)
+            if [ "$HAS_SLAVE" = true ]; then echo "Either --add-slave or --add-organization"; exit 1; fi;
+            SLAVE_SERVICES="${SLAVE_SERVICES} slave-node"
+            HAS_SLAVE=true
+            echo "INFO: slave-node enabled"
+            shift # past argument
+        ;;
+        
+        --add-organization)
+            if [ "$HAS_SLAVE" = true ]; then echo "Either --add-slave or --add-organization"; exit 1; fi;
+            SLAVE_SERVICES="${SLAVE_SERVICES} slave-node slave-api slave-frontend"
+            echo "INFO: slave-node, slave-api, slave-frontend enabled"
+            HAS_SLAVE=true
+            shift # past argument
+        ;;
         --provision)
             WITH_PROVISIONING=true
             shift # past argument
@@ -108,6 +129,7 @@ if [ "$PRUNE_DATA" = true ]; then
     read answer
     if [ "$answer" = "yes" ] || [ "$answer" = "Y" ] || [ "$answer" = "y" ]; then
         sudo rm -r /masterVolume
+        sudo rm -r /slave1Volume
         sudo rm -r /minioVolume
         sudo rm -r /emaildbVolume
     else
@@ -185,8 +207,7 @@ if [ "$IS_FULL" = true ]; then
         COMPOSE_SERVICES="master-node emaildb minio master-api email-service excel-export-service storage-service frontend"
     else
         echo "INFO: Setup full TruBudget environment with provisioning ..."
-        # Empty means all containers
-        COMPOSE_SERVICES=""
+        COMPOSE_SERVICES="master-node emaildb minio master-api email-service excel-export-service storage-service provisioning frontend"
     fi
 else
     if [ "$WITH_PROVISIONING" = false ]; then
@@ -206,11 +227,11 @@ echo "INFO: Pull images from https://hub.docker.com/ ..."
 $COMPOSE pull
 
 echo "INFO: Since images are used, building is not necessary and will be skipped."
-$COMPOSE build
+#$COMPOSE build
 
 # Start docker containers
-echo "INFO: Executing command: $COMPOSE up $LOG_OPTION $COMPOSE_SERVICES $ENABLED_SERVICES"
-$COMPOSE up $LOG_OPTION $COMPOSE_SERVICES $ENABLED_SERVICES
+echo "INFO: Executing command: $COMPOSE up $LOG_OPTION $COMPOSE_SERVICES $ENABLED_SERVICES $SLAVE_SERVICES"
+$COMPOSE up $LOG_OPTION $COMPOSE_SERVICES $ENABLED_SERVICES $SLAVE_SERVICES
 
 if [ "$IS_LOG_ENABLED" = false ]; then
     echo "INFO: Docker container are started without logging"