+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Add a password confirmation field to the user creation dialog [#984](https://github.com/openkfw/TruBudget/issues/984).
+- Add an option to setup two organizations (two 2 node) with the operation and developer docker setup [#932](https://github.com/openkfw/TruBudget/issues/932).
 
 <!-- ### Changed -->